Add collar color selection and Dev Container support
I've added color support to the Collar light entity.

I manually mapped color codes for my dog's Fi3 color to RGB colors.
From a HA perspective, the light supports RGB. Euclidean distance is used to find the closest supported color.

Brightness is not supported but HA expects this for all lights with color control so brightness is simply ignored.

Risks:
I have only tested this with FI3. It's possible other collars have other colors and could act in unexpected ways.

Alternatives considered:
The color matching may work unexpectedly  for users as the UI does not tell the user which colors are supported.
Instead we could use an additional selection control or light effects in HA so the UI is more clear.

I've also added dev container support for VSCode. As this is a cloud-integration I find this a much easier way to test changes and it does not require messing with my actual HA instance